worked on login mostly
all that's left in login is to actually compare the credentials.
app.js: added forbidden files (prefs.html), some comments, and bettered the "autoserve".
in login.js: added result checking, removed the debug alerts, and prepared it for prefs recieving.
error.html: fixed a small thing where the css wasn't loading because of changes to the structure.